Clean code with Codacy rules
